src/applications/microservices/petlistadoptions-py/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.10
botocore==1.40.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
mypy==1.17.1
mypy_extensions==1.1.0
pathspec==0.12.1
prometheus_client==0.22.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
types-psycopg2==2.9.21.20250809
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
